#645 add create button to all features forms
