+++ apps/belis/desktop/src/components/ui/featuresForm/AbzweigdoseForm.tsx
@@ -7,6 +7,7 @@ import { getJWT } from "../../../store/slices/auth";
 import { DokumentItem } from "../DocumentPreview";
 import { getDocumentKey } from "../FilePreview";
 import FeatureFormLayout from "./FeatureFormLayout";
+import { useCreateFeatureDraft } from "../useCreateFeatureDraft";
 import { extractListItem } from "../BelisSidebar";
 import { uploadDraftFiles } from "../../../helper/uploadDraftFiles";
 import { updateDataByClassName } from "../../../helper/apiMethods";
@@ -54,6 +55,7 @@ const AbzweigdoseForm = ({
     null
   );
   const jwt = useSelector(getJWT);
+  const createFeatureDraft = useCreateFeatureDraft();
 
   const handleToggleRemoveDocument = useCallback(
     (key: string) => {
@@ -192,6 +194,7 @@ const AbzweigdoseForm = ({
       onToggleReadOnly={onToggleReadOnly}
       onCancel={onCancel}
       onSave={handleSave}
+      onCreateRelatedDraft={() => createFeatureDraft("abzweigdose")}
       singleColumn
     >
       {null}
