+++ apps/belis/desktop/src/components/ui/featuresForm/MauerlascheForm.tsx
@@ -9,6 +9,7 @@ import { getJWT } from "../../../store/slices/auth";
 import { DokumentItem } from "../DocumentPreview";
 import { getDocumentKey } from "../FilePreview";
 import FeatureFormLayout from "./FeatureFormLayout";
+import { useCreateFeatureDraft } from "../useCreateFeatureDraft";
 import { extractListItem } from "../BelisSidebar";
 import MauerlascheFormFields from "./MauerlascheFormFields";
 import toTitleCase from "../../../helper/toTitleCase";
@@ -81,6 +82,7 @@ const MauerlascheForm = ({
     null
   );
   const jwt = useSelector(getJWT);
+  const createFeatureDraft = useCreateFeatureDraft();
 
   const setFormInstance = useCallback((form: FormInstance) => {
     formRef.current = form;
@@ -252,6 +254,7 @@ const MauerlascheForm = ({
       onToggleReadOnly={onToggleReadOnly}
       onCancel={onCancel}
       onSave={handleSave}
+      onCreateRelatedDraft={() => createFeatureDraft("mauerlasche")}
     >
       <MauerlascheFormFields
         mauerlasche={ml}
